Replace the pdf renderer with react-pdf

+++ packages/react-sdk/src/components/BoardBar/pdfv2/createWhiteboardPdf.ts
@@ -16,7 +16,8 @@
 
 import { WidgetApi } from '@matrix-widget-toolkit/api';
 import { from, fromEvent, map, Observable, switchMap, take, tap } from 'rxjs';
-import { WhiteboardInstance } from '../../../state';
+import { WhiteboardDocumentExport, WhiteboardInstance } from '../../../state';
+import { conv2png } from '../pdf/utils';
 
 export function createWhiteboardPdf(params: {
   whiteboardInstance: WhiteboardInstance;
@@ -30,6 +31,10 @@ export function createWhiteboardPdf(params: {
     const worker = new Worker(new URL('./pdf.worker.ts', import.meta.url), {
       type: 'module',
     });
+    worker.onmessageerror = (e) => {
+      console.error('Worker error', e);
+      worker.terminate();
+    }
 
     // Post the whiteboard instance to the worker and then return the blob that the worker sends back to us when it's done.
     // We must return an observable that emits the blob and then completes.
@@ -37,6 +42,9 @@ export function createWhiteboardPdf(params: {
     // Also note that we need the whiteboardExport which is a promise and the resolved value must be passed to the worker.
     return from(whiteboardExport).pipe(
       switchMap((exportData) => {
+        // Convert all images which are not png or jpeg to png
+        exportData = convertAllImagesToPNG(exportData);
+
         // Pass the whiteboard export data to the worker and wait for the worker to send back the blob.
         // If the worker sends back an error, we should throw it and stop the worker.
         // If the worker sends back the blob, we should return it. We should also stop the worker.
@@ -44,19 +52,28 @@ export function createWhiteboardPdf(params: {
         // Create an observable that emits the blob when the worker sends it back.
         const stream = fromEvent<MessageEvent>(worker, 'message');
 
-        // Note that we cant directly return the stream as it needs to be Observable<Blob> and not Observable<MessageEvent<Blob>
-        worker.postMessage(exportData);
+        if (import.meta.hot) {
+          // Delay the post message to the worker to ensure the worker is ready to receive it.
+          new Promise(resolve => setTimeout(resolve, 2000)).then(() => {
+            console.log('Posting message to worker');
+            worker.postMessage(exportData);
+            return;
+          }).catch((e) => {
+            console.error('Error posting message to worker', e);
+            worker.terminate();
+          });
+        } else {
+          // Note that we cant directly return the stream as it needs to be Observable<Blob> and not Observable<MessageEvent<Blob>
+          console.log('Posting message to worker');
+          worker.postMessage(exportData);
+        }
         return stream;
       }),
       map((e) => e.data),
       take(1),
       tap({
-        unsubscribe: () => {
-          // terminate the worker early
-          worker.terminate();
-        },
-        complete: () => {
-          // terminate the worker when ready
+        finalize: () => {
+          // terminate the worker
           worker.terminate();
         },
       }),
@@ -65,9 +82,42 @@ export function createWhiteboardPdf(params: {
     // Async import pdf.local.ts and thn call renderPDF with the whiteboard export data which is a promise too.
     return from(import('./pdf.local')).pipe(
       switchMap(({ renderPDF }) =>
-        from(whiteboardExport).pipe(switchMap(renderPDF)),
+        from(whiteboardExport).pipe(switchMap((exportData) => {
+          // Convert all images which are not png or jpeg to png
+          exportData = convertAllImagesToPNG(exportData);
+          return renderPDF(exportData);
+        })),
       ),
       take(1),
     );
   }
+}
+
+function convertAllImagesToPNG(exportData: WhiteboardDocumentExport): WhiteboardDocumentExport {
+  // Convert all images which are not png or jpeg to png
+  exportData.whiteboard.slides.forEach((slide) => {
+    slide.elements.forEach((element) => {
+      if (element.type === 'image' && element.mimeType !== 'image/png') {
+        const file = exportData.whiteboard.files?.find((f) => f.mxc === element.mxc);
+        if (file) {
+          const data = conv2png(element, file.data);
+          file.data = data;
+          element.mimeType = 'image/png';
+        }
+      }
+    });
+  });
+
+  // Sanity check in debug mode. Report error if any image is not converted to png or jpeg.
+  if (import.meta.env.MODE === 'development') {
+    exportData.whiteboard.slides.forEach((slide) => {
+      slide.elements.forEach((element) => {
+        if (element.type === 'image' && element.mimeType !== 'image/png') {
+          console.error('Image not converted to png or jpeg', element);
+        }
+      });
+    });
+  }
+
+  return exportData;
 }
